mv Tabs.primitive.tsx -> Tabs.tsx
